requirements.txt
alembic==1.4.2
aniso8601==8.0.0
atomicwrites==1.4.0
attrs==19.3.0
bcrypt==3.1.7
blinker==1.4
boto3==1.14.20
botocore==1.17.20
cffi==1.14.0
click==7.1.2
colorama==0.4.3
dnspython==1.16.0
docutils==0.15.2
email-validator==1.1.1
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Bcrypt==0.7.1
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.4.0
numpy==1.18.5
packaging==20.4
pandas==0.24.2
pathlib2==2.3.5
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pycryptodome==3.9.8
PyJWT==1.4.2
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python2-secrets==1.0.5
pytz==2020.1
s3transfer==0.3.3
secret==0.8
six==1.15.0
SQLAlchemy==1.3.18
tabulate==0.8.7
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.1
zipp==1.2.0

